MNT: Re-rendered with conda-smithy 2.3.2.post.dev52
Re-renders with a dev version of `conda-smithy`, which provides support
for build matrices in CircleCI. While this feature has had some testing,
it has not gone live yet and should not generally be used until it is
released.

Adds in the other NumPy versions to CircleCI and Travis CI matrices.