Switch macOS builds to CircleCI
Travis CI builds are running into log length limits, which are causing
the macOS builds there to fail. It's very hard to "fix" the builds to
stay within this log length limit and arguably not worth the effort.
Given this, simply switch macOS builds to CircleCI where the constraints
are generally more reasonable.