Hackily try to bypass weird Fortran compiler usage
For some reason the `devtoolset` compilers want to use Fortran compilers
with non-Fortran code. This is a hack that attempts to call GCC a
Fortran compiler. Definitely not a real fix and far from ideal. However
this is only needed for the legacy compilers and will be removed as soon
as they are dropped. So this may be sufficient for now.